Lab-1 extracted into a separate module
