+++ CMakeLists.txt
@@ -3,4 +3,8 @@ project(programming)
 
 set(CMAKE_CXX_STANDARD 17)
 
+add_subdirectory(lab-1)
+
 add_executable(programming main.cpp)
+
+target_link_libraries(programming PRIVATE lab-1)
