--- cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-17941af0d22a269c0f42.json
@@ -1,60 +1,0 @@
-{
-	"configurations" : 
-	[
-		{
-			"directories" : 
-			[
-				{
-					"build" : ".",
-					"jsonFile" : "directory-.-Debug-d0094a50bb2071803777.json",
-					"minimumCMakeVersion" : 
-					{
-						"string" : "3.27"
-					},
-					"projectIndex" : 0,
-					"source" : ".",
-					"targetIndexes" : 
-					[
-						0
-					]
-				}
-			],
-			"name" : "Debug",
-			"projects" : 
-			[
-				{
-					"directoryIndexes" : 
-					[
-						0
-					],
-					"name" : "programming",
-					"targetIndexes" : 
-					[
-						0
-					]
-				}
-			],
-			"targets" : 
-			[
-				{
-					"directoryIndex" : 0,
-					"id" : "programming::@6890427a1f51a3e7e1df",
-					"jsonFile" : "target-programming-Debug-90f4a58bda85b49567e2.json",
-					"name" : "programming",
-					"projectIndex" : 0
-				}
-			]
-		}
-	],
-	"kind" : "codemodel",
-	"paths" : 
-	{
-		"build" : "C:/Users/MSI/CLionProjects/nstu/programming/cmake-build-debug",
-		"source" : "C:/Users/MSI/CLionProjects/nstu/programming"
-	},
-	"version" : 
-	{
-		"major" : 2,
-		"minor" : 6
-	}
-}
