+++ cmake-build-debug/CMakeFiles/rules.ninja
@@ -39,6 +39,25 @@ rule CUSTOM_COMMAND
 
 
 #############################################
+# Rule for compiling CXX files.
+
+rule CXX_COMPILER__lab-1_unscanned_Debug
+  depfile = $DEP_FILE
+  deps = gcc
+  command = ${LAUNCHER}${CODE_CHECK}C:\Users\MSI\AppData\Local\Programs\CLion\bin\mingw\bin\g++.exe $DEFINES $INCLUDES $FLAGS -MD -MT $out -MF $DEP_FILE -o $out -c $in
+  description = Building CXX object $out
+
+
+#############################################
+# Rule for linking CXX static library.
+
+rule CXX_STATIC_LIBRARY_LINKER__lab-1_Debug
+  command = cmd.exe /C "$PRE_LINK && C:\Users\MSI\AppData\Local\Programs\CLion\bin\cmake\win\x64\bin\cmake.exe -E rm -f $TARGET_FILE && C:\Users\MSI\AppData\Local\Programs\CLion\bin\mingw\bin\ar.exe qc $TARGET_FILE $LINK_FLAGS $in && C:\Users\MSI\AppData\Local\Programs\CLion\bin\mingw\bin\ranlib.exe $TARGET_FILE && $POST_BUILD"
+  description = Linking CXX static library $TARGET_FILE
+  restat = $RESTAT
+
+
+#############################################
 # Rule for re-running cmake.
 
 rule RERUN_CMAKE
