+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Feb 25 11:56 RTZ 6 (çèìà)
+Start testing: Feb 25 12:24 RTZ 6 (çèìà)
 ----------------------------------------------------------
-End testing: Feb 25 11:56 RTZ 6 (çèìà)
+End testing: Feb 25 12:24 RTZ 6 (çèìà)
