+++ lab-1/lab_1.h
@@ -1,0 +1,10 @@
+//
+// Created by MSI on 25.02.2024.
+//
+
+#ifndef PROGRAMMING_LAB_1_H
+#define PROGRAMMING_LAB_1_H
+
+int show();
+
+#endif //PROGRAMMING_LAB_1_H
